use blob url for videoMeta
